add context and op
op will be the backbone of the whole framework -- the idea is to weave
multiple ops together based on their inputs and ouputs, and restrict to
operate only what they declare. they don't need to drop things
explicitly as the framework will be able to do that for them as needed.

Change-Id: Id000000043cc1212a47ac09e7387cb670b29c38b